Airspeed Selector/Validator: add airspeed confidence value to airspeed_validated topic
The airspeed confidence is a metric from the airspeed driver, indicating
from 0 to 1 how confident the driver is about the current data coming
from the sensor. This commit routes through the confidence value from the
driver to the output of the airspeed selector, to be used by e.g. the
airspeed prearm check.

Signed-off-by: Silvan <silvan.fuhrer@hotmail.com>